Arrumei foi tudo nessa joça...
