+++ frontend/src/components/Main.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Route } from "react-router-dom";
-import Home from "./Home";
+
+import { Home } from "./Home";
 
 function Main() {
   return (
