+++ frontend/src/services/requestsDb.js
@@ -1,7 +1,5 @@
 import axios from "axios";
 
-const apiDb = axios.create({
+export const api = axios.create({
   baseURL: "http://localhost:3001/",
 });
-
-export default apiDb;
